requirements.txt
streamlit>=1.30
pytest>=7.0
anthropic>=0.25
python-dotenv>=1.0
